pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "基于 LangGraph 的 AI Agent 服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.1",
    "httpx>=0.28.1",
    "langchain-core>=1.2.0",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.0",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration_real: tests that require real API keys",
]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
pretty = true
exclude = ["tests/", "examples/", "scripts/"]

[[tool.mypy.overrides]]
module = [
    "langgraph.*",
    "langsmith.*",
    "langchain_core.*",
    "langchain_openai.*",
    "tiktoken.*",
    "pydantic_settings.*",
]
ignore_missing_imports = true
